fix(stats-carousel): move component to patterns directory
