+++ src/components/patterns/statsCarousel/defs.ts
