+++ src/components/patterns/statsCarousel/index.ts
