+++ src/components/patterns/statsCarousel/statsCarousel.scss
@@ -1,4 +1,4 @@
-@use '../../scss/utility/component-palettes.scss';
+@use '../../../scss/utility/component-palettes.scss';
 @use '~@kyndryl-design-system/shidoka-foundation/scss/variables/breakpoints.scss';
 @use '~@kyndryl-design-system/shidoka-foundation/scss/utility/grid.scss';
 @use '~@kyndryl-design-system/shidoka-foundation/scss/utility/typography.scss';
