+++ src/components/patterns/statsCarousel/statsCarousel.ts
