+++ src/components/patterns/statsCarousel/statsCarouselSlide.scss
