+++ src/components/patterns/statsCarousel/statsCarouselSlide.ts
